+++ tests/support/FunctionalTester.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace davidhirtz\yii2\cms\tests\support;
 
+use davidhirtz\yii2\cms\models\Entry;
+use Yii;
+
 /**
  * Inherited Methods
  * @method void wantTo($text)
@@ -22,4 +25,18 @@ namespace davidhirtz\yii2\cms\tests\support;
 class FunctionalTester extends \Codeception\Actor
 {
     use _generated\FunctionalTesterActions;
+
+    public function amOnDraftSubdomain(): void
+    {
+        $httpHost = parse_url(Yii::$app->getUrlManager()->getDraftHostInfo(), PHP_URL_HOST);
+        $this->haveServerParameter('HTTP_HOST', $httpHost);
+    }
+
+
+    public function grabEntryFixture(string $key): Entry
+    {
+        /** @var Entry $entry */
+        $entry = $this->grabFixture('entries', $key);
+        return $entry;
+    }
 }