+++ tests/support/fixtures/EntryFixture.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace davidhirtz\yii2\cms\tests\support\fixtures;
 
 use davidhirtz\yii2\cms\tests\data\models\TestEntry;