+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.4.0 (May 26, 2025)
+
+- Requires PHP 8.3+
+
+# 2.3.14 (May 22, 2025)
+
+- Enhanced `SiteController::getQuery()` to always apply `EntryQuery::addSelectI18nSlugTargetAttributes()`
+- Fixed `FrontendLink` section draft status (Issue #18)
+
 # 2.3.13 (May 16, 2025)
 
 - Added `SlugAttributeTrait::$slugLowercase` and `SlugAttributeTrait::$slugReplacement`